Add Analytics module and consolidate admin business logic
Create app/Modules/Analytics/ with three services:
- GameStatsService: team popularity, formations, mentality, season progress
- ActivationFunnelService: cohort-based funnel with drop-off calculations
- DashboardStatsService: summary counts (users, games, 7-day trends)

All admin view classes are now thin orchestrators that delegate to
these services. Remove the temporary app/Services/ directory.
